Fixed bench changing bug
Bench: 3380203

+++ src/engine/search/search.cc
@@ -577,8 +577,8 @@ Score Search::PVSearch(int depth,
       int reduction = tables::kLateMoveReduction[is_quiet][depth][moves_seen];
       reduction += !in_pv_node;
       reduction += cut_node;
-      reduction -=
-          is_quiet * history_.GetQuietMoveScore(move, stack) / lmr_hist_div;
+      reduction -= is_quiet * history_.GetQuietMoveScore(move, stack) /
+                   static_cast<int>(lmr_hist_div);
       reduction -= state.InCheck();
 
       // Ensure the reduction doesn't give us a depth below 0
